+++ nmrs/src/core/mod.rs
@@ -9,6 +9,7 @@ pub(crate) mod bluetooth;
 pub(crate) mod connection;
 pub(crate) mod connection_settings;
 pub(crate) mod connectivity;
+pub(crate) mod custom_connection;
 pub(crate) mod device;
 pub(crate) mod ovpn_parser;
 pub(crate) mod rfkill;